Added a new Level Generation framework, and data/config files.
Ripped out pretty much all of world.cpp and made a LevelFactory class
that takes care of building an entire level.
So far the level generation is very fast and nice. It's also very simple
for now, but all the basic building blocks are there so it can and will
be expanded.
I also added jsoncpp to read json files for config and data.